CodeHygiene.Accounts.User (Code Hygiene example app v0.1.0)

The user schema

Functions

confirm_changeset(user)

Confirms the account by setting confirmed_at.

email_changeset(user, attrs)

A user changeset for changing the email.

It requires the email to change otherwise an error is added.

password_changeset(user, attrs, opts \\ [])

A user changeset for changing the password.

Options

:hash_password - Hashes the password so it can be stored securely in the database and ensures the password field is cleared to prevent leaks in the logs. If password hashing is not needed and clearing the password field is not desired (like when using this changeset for validations on a LiveView form), this option can be set to false. Defaults to true.

registration_changeset(user, attrs, opts \\ [])

A user changeset for registration.

It is important to validate the length of both email and password. Otherwise databases may truncate the email without warnings, which could lead to unpredictable or insecure behaviour. Long passwords may also be very expensive to hash for certain algorithms.

Options

:hash_password - Hashes the password so it can be stored securely in the database and ensures the password field is cleared to prevent leaks in the logs. If password hashing is not needed and clearing the password field is not desired (like when using this changeset for validations on a LiveView form), this option can be set to false. Defaults to true.

valid_password?(arg1, password)

Verifies the password.

If there is no user or the user doesn't have a password, we call Bcrypt.no_user_verify/0 to avoid timing attacks.

validate_current_password(changeset, password)

Validates the current password otherwise adds an error to the changeset.